OSX: Move winit::Window to platform::Window
