+++ src/api/mod.rs
@@ -1,6 +1,5 @@
 pub mod android;
 pub mod caca;
-pub mod cocoa;
 pub mod dlopen;
 pub mod egl;
 pub mod emscripten;
