+++ src/platform/macos/helpers.rs
